moved most of CLIUtils into run and convert packages
