+++ vcell-cli/src/main/java/org/vcell/cli/CLIStandalone.java
@@ -1,6 +1,8 @@
 package org.vcell.cli;
 
 
+import org.vcell.cli.run.ExecuteCommand;
+import org.vcell.cli.vcml.ConvertCommand;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Spec;
