+++ vcell-cli/src/main/java/org/vcell/cli/run/ExecuteCommand.java
@@ -1,4 +1,4 @@
-package org.vcell.cli;
+package org.vcell.cli.run;
 
 import cbit.vcell.parser.ExpressionException;
 import cbit.vcell.resource.OperatingSystemInfo;
@@ -6,6 +6,9 @@ import cbit.vcell.resource.PropertyLoader;
 import cbit.vcell.solver.ode.ODESolverResultSet;
 import cbit.vcell.xml.ExternalDocInfo;
 import org.jlibsedml.*;
+import org.vcell.cli.CLIPythonManager;
+import org.vcell.cli.CLIUtils;
+import org.vcell.cli.PythonStreamException;
 import org.vcell.cli.vcml.VCMLHandler;
 import org.vcell.util.FileUtils;
 import org.vcell.util.GenericExtensionFilter;
@@ -27,7 +30,7 @@ import java.util.Map;
 import java.util.concurrent.Callable;
 
 @Command(name = "execute")
-class ExecuteCommand implements Callable<Integer> {
+public class ExecuteCommand implements Callable<Integer> {
 
     @Option(names = { "-i", "--inputFilePath" })
     private File inputFilePath;
@@ -142,7 +145,7 @@ class ExecuteCommand implements Callable<Integer> {
         File outDirForCurrentVcml = new File(Paths.get(outputDir.getAbsolutePath(), vcmlName).toString());
 
         try {
-            CLIUtils.removeAndMakeDirs(outDirForCurrentVcml);
+            RunUtils.removeAndMakeDirs(outDirForCurrentVcml);
         } catch (Exception e) {
             System.err.println("Error in creating required directories: " + e.getMessage());
             e.printStackTrace(System.err);
@@ -156,8 +159,8 @@ class ExecuteCommand implements Callable<Integer> {
 
             for (String simName : resultsHash.keySet()) {
                 String CSVFilePath = Paths.get(outDirForCurrentVcml.toString(), simName + ".csv").toString();
-                CLIUtils.createCSVFromODEResultSet(resultsHash.get(simName), new File(CSVFilePath));
-                CLIUtils.transposeVcmlCsv(CSVFilePath);
+                RunUtils.createCSVFromODEResultSet(resultsHash.get(simName), new File(CSVFilePath));
+                RunUtils.transposeVcmlCsv(CSVFilePath);
             }
         } catch (IOException e) {
             System.err.println("IOException while processing VCML " + vcmlFile.getName());
@@ -207,7 +210,7 @@ class ExecuteCommand implements Callable<Integer> {
         long startTimeOmex = System.currentTimeMillis();
 
         System.out.println("VCell CLI input archive " + inputFileName);
-        CLIUtils.drawBreakLine("-", 100);
+        RunUtils.drawBreakLine("-", 100);
         try {
             sedmlPath2d3d = Paths.get(outputDir, "temp");
             omexHandler = new OmexHandler(inputFileName, outputDir);
@@ -224,8 +227,8 @@ class ExecuteCommand implements Callable<Integer> {
             throw new Exception(error);
         }
 
-        CLIUtils.removeAndMakeDirs(new File(outputDir));
-        CLIUtils.generateStatusYaml(inputFileName, outputDir);    // generate Status YAML
+        RunUtils.removeAndMakeDirs(new File(outputDir));
+        RunUtils.generateStatusYaml(inputFileName, outputDir);    // generate Status YAML
 
         // from here on, we need to collect errors, since some subtasks may succeed while other do not
         // we now have the log file created, so that we also have a place to put them
@@ -248,7 +251,7 @@ class ExecuteCommand implements Callable<Integer> {
             File outDirForCurrentSedml = new File(omexHandler.getOutputPathFromSedml(sedmlLocation));
 
             try {
-                CLIUtils.removeAndMakeDirs(outDirForCurrentSedml);
+                RunUtils.removeAndMakeDirs(outDirForCurrentSedml);
 
                 SedML sedmlFromOmex = Libsedml.readDocument(completeSedmlPath).getSedMLModel();
 
@@ -280,10 +283,10 @@ class ExecuteCommand implements Callable<Integer> {
                 String str = "Successful translation of SED-ML file";
                 logDocumentMessage += str + ". ";
                 System.out.println(str + " : " + sedmlName);
-                CLIUtils.drawBreakLine("-", 100);
+                RunUtils.drawBreakLine("-", 100);
 
                 // For appending data for SED Plot2D and Plot3d to HDF5 files following a temp convention
-                CLIUtils.genSedmlForSed2DAnd3D(inputFileName, outputDir);
+                PythonCalls.genSedmlForSed2DAnd3D(inputFileName, outputDir);
                 // SED-ML file generated by python VCell_cli_util
                 sedmlPathwith2dand3d = new File(String.valueOf(sedmlPath2d3d), "simulation_" + sedmlName);
                 Path path = Paths.get(sedmlPathwith2dand3d.getAbsolutePath());
@@ -304,8 +307,8 @@ class ExecuteCommand implements Callable<Integer> {
                 String prefix = "SED-ML processing for " + sedmlLocation + " failed with error: ";
                 logDocumentError = prefix + e.getMessage();
                 String type = e.getClass().getSimpleName();
-                CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-                CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+                PythonCalls.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+                PythonCalls.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 CLIUtils.writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError, bForceLogFiles);
                 writeDetailedResultList(outputBaseDir, bioModelBaseName + "," + sedmlName + "," + logDocumentError, bForceLogFiles);
 
@@ -313,7 +316,7 @@ class ExecuteCommand implements Callable<Integer> {
                 e.printStackTrace(System.err);
                 somethingFailed = true;
                 oneSedmlDocumentFailed = true;
-                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, CLIUtils.Status.FAILED, outputDir);
+                PythonCalls.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 continue;
             }
 
@@ -369,7 +372,7 @@ class ExecuteCommand implements Callable<Integer> {
                 }
                 logDocumentMessage += "Generating outputs... ";
 
-                reportsHash = CLIUtils.generateReportsAsCSV(sedml, resultsHash, outDirForCurrentSedml, outputDir, sedmlLocation);
+                reportsHash = RunUtils.generateReportsAsCSV(sedml, resultsHash, outDirForCurrentSedml, outputDir, sedmlLocation);
 
                 if (reportsHash == null || reportsHash.size() == 0) {
                     oneSedmlDocumentFailed = true;
@@ -387,9 +390,9 @@ class ExecuteCommand implements Callable<Integer> {
                 }
 
                 logDocumentMessage += "Generating HDF5 file... ";
-                String idNamePlotsMap = CLIUtils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
+                String idNamePlotsMap = RunUtils.generateIdNamePlotsMap(sedml, outDirForCurrentSedml);
 
-                CLIUtils.execPlotOutputSedDoc(inputFileName, idNamePlotsMap, outputDir);                            // create the HDF5 file
+                PythonCalls.execPlotOutputSedDoc(inputFileName, idNamePlotsMap, outputDir);                            // create the HDF5 file
 
                 if (!containsExtension(outputDir, "h5")) {
                     oneSedmlDocumentFailed = true;
@@ -406,10 +409,10 @@ class ExecuteCommand implements Callable<Integer> {
                 oneSedmlDocumentFailed = true;
                 logDocumentError += e.getMessage();
                 String type = e.getClass().getSimpleName();
-                CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-                CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+                PythonCalls.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+                PythonCalls.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 CLIUtils.writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError, bForceLogFiles);
-                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, CLIUtils.Status.FAILED, outputDir);
+                PythonCalls.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 org.apache.commons.io.FileUtils.deleteDirectory(new File(String.valueOf(sedmlPath2d3d)));    // removing temp path generated from python
                 continue;
             }
@@ -418,19 +421,19 @@ class ExecuteCommand implements Callable<Integer> {
                 Exception e = new RuntimeException("Failure executing the sed document. ");
                 logDocumentError += e.getMessage();
                 String type = e.getClass().getSimpleName();
-                CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-                CLIUtils.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
+                PythonCalls.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+                PythonCalls.setExceptionMessage(sedmlLocation, sedmlName, outputDir, "sedml", type, logDocumentError);
                 CLIUtils.writeDetailedErrorList(outputBaseDir, bioModelBaseName + ",  doc:    " + type + ": " + logDocumentError, bForceLogFiles);
-                CLIUtils.updateSedmlDocStatusYml(sedmlLocation, CLIUtils.Status.FAILED, outputDir);
+                PythonCalls.updateSedmlDocStatusYml(sedmlLocation, Status.FAILED, outputDir);
                 org.apache.commons.io.FileUtils.deleteDirectory(new File(String.valueOf(sedmlPath2d3d)));    // removing temp path generated from python
                 continue;
             }
             org.apache.commons.io.FileUtils.deleteDirectory(new File(String.valueOf(sedmlPath2d3d)));    // removing temp path generated from python
 
             // archiving res files
-            CLIUtils.zipResFiles(new File(outputDir));
-            CLIUtils.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
-            CLIUtils.updateSedmlDocStatusYml(sedmlLocation, CLIUtils.Status.SUCCEEDED, outputDir);
+            RunUtils.zipResFiles(new File(outputDir));
+            PythonCalls.setOutputMessage(sedmlLocation, sedmlName, outputDir, "sedml", logDocumentMessage);
+            PythonCalls.updateSedmlDocStatusYml(sedmlLocation, Status.SUCCEEDED, outputDir);
         }
 
         omexHandler.deleteExtractedOmex();
@@ -448,16 +451,16 @@ class ExecuteCommand implements Callable<Integer> {
             if (oneSedmlDocumentSucceeded) {        // some succeeded, some failed
                 error = " At least one document in this archive failed to execute";
             }
-            CLIUtils.updateOmexStatusYml(CLIUtils.Status.FAILED, outputDir, duration + "");
+            PythonCalls.updateOmexStatusYml(Status.FAILED, outputDir, duration + "");
             System.err.println(error);
             logOmexMessage += error;
             writeErrorList(outputBaseDir, bioModelBaseName, bForceLogFiles);
         } else {
-            CLIUtils.updateOmexStatusYml(CLIUtils.Status.SUCCEEDED, outputDir, duration + "");
+            PythonCalls.updateOmexStatusYml(Status.SUCCEEDED, outputDir, duration + "");
             writeFullSuccessList(outputBaseDir, bioModelBaseName, bForceLogFiles);
             logOmexMessage += " Done";
         }
-        CLIUtils.setOutputMessage("null", "null", outputDir, "omex", logOmexMessage);
+        PythonCalls.setOutputMessage("null", "null", outputDir, "omex", logOmexMessage);
 
     }
