+++ vcell-cli/src/main/java/org/vcell/cli/run/Status.java
@@ -1,0 +1,17 @@
+package org.vcell.cli.run;
+
+public enum Status {
+    RUNNING("Simulation still running"),
+    SKIPPED("Simulation skipped"),
+    SUCCEEDED("Simulation succeeded"),
+    FAILED("Simulation failed"),
+    ABORTED("Simulation aborted");
+
+    private final String description;
+
+    Status(String desc) {
+        this.description = desc;
+    }
+
+
+}
